Finish modularizing code + begin working on manual citations
